reporting the tuning parameters
The lsa_function is barrowed from the RobMixReg package and modified to  report lambda values. So this version does not depend on the RobMixReg package, but directly use lars  package instead. The lsa function is extended to support kmcure fit object as a possible input similar to lm/glm/coxph fit objects.